requirements.txt
fastapi
uvicorn[standard]
jinja2
python-dotenv
python-telegram-bot
requests
yookassa
